# filmbox-api/openapi/filmbox/filmboxapi.yaml
swagger: "2.0"
info:
  description: "New user local api."
  version: "1.0.0"
  title: "Swagger Filmbox"
basePath: "/v1"
tags:
  - name: "movie"
    description: "Operations on movie"
schemes:
  - "http"
paths:
  "/movie":
    get:
      tags:
        - "movie"
      summary: "Find all movies"
      operationId: "findAll"
      responses:
        '200':
          'description': 'OK'
          'schema':
            'type': array
            'items':
              $ref: "#/definitions/MovieDto"
        '400':
          'description': 'Bad request'
        '404':
          'description': 'Objects not found'
        '500':
          'description': 'Local server issue'
    post:
      tags:
        - "movie"
      summary: "Create new movie"
      operationId: "create"
      parameters:
        - name: movie
          in: body
          required: true
          "schema":
            $ref: "#/definitions/MovieDto"
      responses:
        '200':
          'description': 'OK'
          'schema':
            $ref: "#/definitions/MovieDto"
        '400':
          'description': 'Bad request'
        '404':
          'description': 'Objects not found'
        '500':
          'description': 'Local server issue'
    put:
      tags:
        - "movie"
      summary: "Update movie"
      operationId: "update"
      parameters:
        - name: movie
          in: body
          required: true
          "schema":
            $ref: "#/definitions/MovieDto"
      responses:
        '200':
          'description': 'OK'
          'schema':
            $ref: "#/definitions/MovieDto"
        '400':
          'description': 'Bad request'
        '404':
          'description': 'Objects not found'
        '500':
          'description': 'Local server issue'
  "/movie/{id}":
    get:
      tags:
        - "movie"
      summary: "Find specific movie by id"
      operationId: "findById"
      parameters:
        - name: id
          in: path
          type: number
          required: true
          description: "Searching movies by specific id"
      responses:
        '200':
          'description': 'OK'
          'schema':
            'type': object
            'items':
              $ref: "#/definitions/MovieDto"
        '400':
          'description': 'Bad request'
        '404':
          'description': 'Objects not found'
        '500':
          'description': 'Local server issue'
  "/movie/search":
    get:
      tags:
        - "movie"
      summary: "Find movies by text, returns all with query name"
      operationId: "findAllByNameContaining"
      parameters:
        - name: text
          in: query
          description: Searching movies by text
          required: false
          type: string
      responses:
        '200':
          'description': 'OK'
          'schema':
            'type': array
            'items':
              $ref: "#/definitions/MovieDto"
        '400':
          'description': 'Bad request'
        '404':
          'description': 'Objects not found'
        '500':
          'description': 'Local server issue'
definitions:
  MovieDto:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Film id"
      name:
        type: "string"
        description: "Film name"
      description:
        type: "string"
        description: "Film description"
      worldReleaseDate:
        type: "string"
        description: "Release date"
      polishReleaseDate:
        type: "string"
        description: "Release date in Poland"